updated task 2

+++ 0x1A-hash_tables/2-key_index.c
@@ -11,7 +11,7 @@
  * Description: Uses the djb2 algorithm.
  */
 
-unsigned long int key_index(const unsigned char *key, unsigned int size)
+unsigned long int key_index(const unsigned char *key, unsigned long int size)
 {
 	return (hash_djb2(key) % size);
 }
